move some more bits into the k8s helper
